# akkor-api/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Akkor website API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/login": {
      "post": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "email and password",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "MyPasswordOk!"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "all of the below body",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                },
                "pseudo": {
                  "type": "string",
                  "example": "MyPseudo"
                },
                "password": {
                  "type": "string",
                  "example": "MyPasswordOk!"
                },
                "role": {
                  "type": "string",
                  "example": "admin"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "at least one of the variables in the body (ex : if you only want to update the role, you can just send the role in the body).",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                },
                "pseudo": {
                  "type": "string",
                  "example": "MyPseudo"
                },
                "password": {
                  "type": "string",
                  "example": "MyPasswordOk!"
                },
                "role": {
                  "type": "string",
                  "example": "admin"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/hotels/": {
      "get": {
        "tags": [
          "Hotel managing"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Hotel managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "role",
            "in": "header",
            "type": "string"
          },
          {
            "name": "hotel",
            "in": "body",
            "description": "All of the body below.",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "HotelName"
                },
                "location": {
                  "type": "string",
                  "example": "Paris"
                },
                "description": {
                  "type": "string",
                  "example": "Just a hotel"
                },
                "pictureList": {
                  "type": "array",
                  "example": [
                    "blob1",
                    "blob2"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/hotels/{id}": {
      "get": {
        "tags": [
          "Hotel managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Hotel managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "role",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Hotel managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "role",
            "in": "header",
            "type": "string"
          },
          {
            "name": "hotel",
            "in": "body",
            "description": "at least one of the variables in the body (ex : if you only want to update the name, you can just send the name in the body).",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "HotelName"
                },
                "location": {
                  "type": "string",
                  "example": "Paris"
                },
                "description": {
                  "type": "string",
                  "example": "Just a hotel"
                },
                "pictureList": {
                  "type": "array",
                  "example": [
                    "blob1",
                    "blob2"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bookings/": {
      "get": {
        "tags": [
          "User managing"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Booking managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "booking",
            "in": "body",
            "description": "All of the body below.",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "hotelID": {
                  "type": "string",
                  "example": "g5sg4sh4s9g4s9"
                },
                "from": {
                  "type": "string",
                  "example": "02/08/2023"
                },
                "to": {
                  "type": "string",
                  "example": "02/09/2023"
                },
                "userId": {
                  "type": "string",
                  "example": "dh4h64dh4d6"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": [
          "Booking managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Booking managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Booking managing"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "booking",
            "in": "body",
            "description": "at least one of the variables in the body (ex : if you only want to update from, you can just send from in the body).",
            "required": true,
            "format": "json",
            "schema": {
              "type": "object",
              "properties": {
                "hotelID": {
                  "type": "string",
                  "example": "g5sg4sh4s9g4s9"
                },
                "from": {
                  "type": "string",
                  "example": "02/08/2023"
                },
                "to": {
                  "type": "string",
                  "example": "02/09/2023"
                },
                "userId": {
                  "type": "string",
                  "example": "dh4h64dh4d6"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}